Remove time-related classes
The classes were not designed to integrate with java.time and had more
sense before java.time existed. However, there are better means now in
Java runtime libraries to deal with time quantities, so instead of trying
to bring the implementation into usable state, it has been dropped.

Be happy with java.time!