Rework the fluent collection hierarchy
Abandoned (for now at least) the concept of split interfaces (extensions
and the combined one). The major reason is reducing the complexity and
increasing natural type compatibility, so that all Fluent- interfaces
could follow the original type hierarchy.

Before this move, a reason against it was the problem with common methods
like map() or remap() which clash in the inherited interfaces, preventing
the mentioned type compatibility. This has been solved by making more
methods with distinct signatures (e.g., withCollection()) and reusing
Stream as the replacement for Source (which loses its sense therefore).

One disadvantage is duplication of many default methods in order to have
correct return types and overriding possibly existing default methods in
the parent interfaces. On the other hand, the combination can reduce the
duplication for the inherited types and provides an effective and yet
augmented forwarding implementation.