Work around a compiler failure
The compiler may fail when compiling Cloneables::prototype when the
cloning function is passed as a lambda directly. The analysis of the
throwable exceptions fails to realize that the actual exception is a
RuntimeException.

When the line is split and the lambda is constructed separately, the
compiler accepts the input.