Optimize ByteSequence views for non-constant data
The views are used even internally for viewing the result of a builder.
The results then are safe and should behave as well as ByteContainer, but
actually suffer from not-caching the hash code. In order to enable that,
the classes got alternative construction methods for use with non-constant
data. Although it creates a potential risk, other related risks existed
before anyway (the synchronization needs). Therefore this relaxation does
not change much.